refactor: extract PathMaterialization types into leaf core module
The PathMaterialization type cluster (PathMaterializationMode / Plan / Entry /
Projection / PathRemap, their impls, and the PATH_MATERIALIZATION_* constants)
lived in runner_execution_envelope, forcing the lab-contract type layer
(command_contract::lab::handoff) to depend on that whole envelope module for a
single PathMaterializationPlan field.

The cluster is fully self-contained (no external crate deps). Move it into a new
leaf core::path_materialization module and re-export it from
runner_execution_envelope to keep all existing call sites stable. Point
lab/handoff.rs at the leaf module.

Result: all three lab-contract files now depend only on leaf / near-leaf modules
(types.rs: none; workload.rs: agent_task_schedule, notification_route,
secret_env_plan; handoff.rs: path_materialization). The deep couplings are gone.

Third cut (3b) toward breaking the core <-> command_contract cycle.
